fix(tabs): current tab still active if selected tab does not have a root
The current selected tab should NOT be deselected (i.e. detached from change detection) if the selected tab does not have a root
ie. a tab that acts as a button to open a modal, logout etc.

Lifecycle events should not be dispatched either. Right now we are dispatching willLeave/willEnter always (this is a bug).

Closes #9392. Closes #9811. Closes #9392.